+++ armadillo/src/test/java/org/molgenis/armadillo/TestHelpers.java
@@ -1,0 +1,26 @@
+package org.molgenis.armadillo;
+
+import java.lang.reflect.Field;
+
+public class TestHelpers {
+  public static void setField(Object target, String fieldName, Object value) throws Exception {
+    Field f = findField(target.getClass(), fieldName);
+    f.setAccessible(true);
+    f.set(target, value);
+  }
+
+  public static Object getField(Object target, String fieldName) throws Exception {
+    Field f = findField(target.getClass(), fieldName);
+    f.setAccessible(true);
+    return f.get(target);
+  }
+
+  public static Field findField(Class<?> clazz, String name) throws NoSuchFieldException {
+    try {
+      return clazz.getDeclaredField(name);
+    } catch (NoSuchFieldException e) {
+      if (clazz.getSuperclass() != null) return findField(clazz.getSuperclass(), name);
+      throw e;
+    }
+  }
+}
